refactor(envelope): split inner.rs into pure types + async signing (#1446)
Prerequisite for scp-protocol crate extraction. Separates pure sync
protocol logic (types, verification, validation) from async signing
(create_inner_envelope) to enable clean extraction boundary.

Closes #0 (no issue — prerequisite for #1446)
